requirements.txt
python-dotenv==1.0.1
google-generativeai==0.8.3
arize-phoenix==7.7.1
llama-index-core==0.12.11
llama-index-llms-gemini==0.4.3
llama-index-utils-workflow==0.3.0
llama-index-readers-file==0.4.3
llama-index-postprocessor-cohere-rerank==0.3.0
llama-index-embeddings-gemini==0.3.1